Added job level calculation, objective completion messages and sounds

+++ src/main/java/me/exyin/jobQuests/model/player/PlayerJob.java
@@ -13,6 +13,6 @@ import java.util.List;
 @AllArgsConstructor
 public class PlayerJob {
     private String jobId;
-    private long xp;
+    private double xp;
     private List<PlayerQuest> playerQuests;
 }
